Restructure SCSS files to use components, rename 'editor' to 'code-example'
