+++ assets/index.css
@@ -1,0 +1,1 @@
+pre{background-color:#333;color:#e5e5e5}pre .token.punctuation,pre .token.operator{color:#e5e5e5}pre .token.tag,pre .token.function{color:#fb52b0}pre .token.attr-name,pre .token.keyword{color:#cbfd66}pre .token.attr-value,pre .token.string{color:#6fcff8}.code-example .output pre{background-color:#fff;color:#555}.code-example .output pre .token.punctuation,.code-example .output pre .token.operator{color:#555}.code-example .output pre .token.tag,.code-example .output pre .token.function{color:#ac0060}.code-example .output pre .token.attr-name,.code-example .output pre .token.keyword{color:#71a013}.code-example .output pre .token.attr-value,.code-example .output pre .token.string{color:#087eb0}body{margin:0;font-family:"Sarabun",Trebuchet MS,Helvetica,sans-serif;font-size:1.125rem}*,*::before,*::after{box-sizing:border-box}*+h2{margin-top:5rem}h2{font-size:1.5rem;font-weight:300;letter-spacing:2px;text-transform:uppercase;color:#889096}p{line-height:1.4}blockquote{margin-left:1rem;border-left:3px solid #aaa;padding-left:.5rem}blockquote+p{margin-left:1rem;font-size:.9rem}img{max-width:100%}pre{display:block;padding:1rem}pre{font-family:"Fira Code",monospace;overflow-x:auto}.container{padding:3rem 1rem}.container>*{margin-left:auto;margin-right:auto;max-width:40rem}.logo{max-width:200px;max-height:200px;width:100%;margin:auto}.title-block{background-color:#36454f;color:#fff}.title-block .container{display:flex;flex-wrap:wrap;margin:0 auto;max-width:40rem}.title-block .title-area{flex-basis:5rem;flex-grow:1}.title-block h1{font-size:3rem;font-family:"Fira Code",monospace}.title-block p{font-size:1.25rem;color:#00f1ea}.code-example{max-width:60rem;margin:3rem auto;display:flex;flex-wrap:wrap;border:1px solid #000;background-color:#333;color:#e5e5e5;box-shadow:0 10px 15px -3px rgba(0,0,0,.1),0 4px 6px -2px rgba(0,0,0,.05)}.code-example pre{margin:0}.code-example .input,.code-example .output{flex-grow:1;margin:0;padding-top:.5rem;overflow-x:auto}.code-example .input::before,.code-example .output::before{content:attr(aria-label);display:block;font-size:.875rem;margin-left:1rem;font-family:"Fira Code",monospace;text-transform:uppercase}.code-example .output{background-color:#fff;color:#555}
