+++ assets/manifest.json
@@ -1,1 +1,1 @@
-{"code.css":"/assets/code.css","fira-code-all-400.woff":"/assets/fonts/fira-code-all-400.woff","fira-code-all-400.woff2":"/assets/fonts/fira-code-all-400.woff2","fonts.css":"/assets/fonts.css","sarabun-all-400-italic.woff":"/assets/fonts/sarabun-all-400-italic.woff","sarabun-all-400-italic.woff2":"/assets/fonts/sarabun-all-400-italic.woff2","sarabun-all-400.woff":"/assets/fonts/sarabun-all-400.woff","sarabun-all-400.woff2":"/assets/fonts/sarabun-all-400.woff2","sarabun-all-700-italic.woff":"/assets/fonts/sarabun-all-700-italic.woff","sarabun-all-700-italic.woff2":"/assets/fonts/sarabun-all-700-italic.woff2","sarabun-all-700.woff":"/assets/fonts/sarabun-all-700.woff","sarabun-all-700.woff2":"/assets/fonts/sarabun-all-700.woff2","views.js":"/dist/views.js"}
+{"fira-code-all-400.woff":"/assets/fonts/fira-code-all-400.woff","fira-code-all-400.woff2":"/assets/fonts/fira-code-all-400.woff2","fonts.css":"/assets/fonts.css","index.css":"/assets/index.css","sarabun-all-400-italic.woff":"/assets/fonts/sarabun-all-400-italic.woff","sarabun-all-400-italic.woff2":"/assets/fonts/sarabun-all-400-italic.woff2","sarabun-all-400.woff":"/assets/fonts/sarabun-all-400.woff","sarabun-all-400.woff2":"/assets/fonts/sarabun-all-400.woff2","sarabun-all-700-italic.woff":"/assets/fonts/sarabun-all-700-italic.woff","sarabun-all-700-italic.woff2":"/assets/fonts/sarabun-all-700-italic.woff2","sarabun-all-700.woff":"/assets/fonts/sarabun-all-700.woff","sarabun-all-700.woff2":"/assets/fonts/sarabun-all-700.woff2","views.js":"/dist/views.js"}
