+++ lib/styles/index.scss
@@ -1,0 +1,59 @@
+@import "./custom-prism";
+
+body {
+	margin: 0;
+	font-family: 'Sarabun', Trebuchet MS, Helvetica, sans-serif;
+	font-size: 1.125rem;
+}
+
+*, *::before, *::after {
+	box-sizing: border-box;
+}
+
+* + h2 {
+	margin-top: 5rem;
+}
+
+h2 {
+	font-size: 1.5rem;
+	font-weight: 300;
+	letter-spacing: 2px;
+	text-transform: uppercase;
+	color: hsl(205.7, 6.3%, 56.1%);
+}
+
+p {
+	line-height: 1.4;
+}
+
+blockquote {
+	margin-left: 1rem;
+	border-left: 3px solid #AAA;
+	padding-left: 0.5rem;
+}
+
+blockquote + p {
+	margin-left: 1rem;
+	font-size: 0.9rem;
+}
+
+img {
+	max-width: 100%;
+}
+
+pre {
+	display: block;
+	padding: 1rem;
+}
+
+pre {
+	@extend %custom-prism-dark-mode;
+	font-family: 'Fira Code', monospace;
+	overflow-x: auto;
+}
+
+/* components */
+@import "../../views/components/container/container";
+@import "../../views/components/logo/logo";
+@import "../../views/components/title_block/title_block";
+@import "../../views/components/code_example/code_example";
