+++ rationale.html
@@ -1,88 +1,5 @@
 <!DOCTYPE html>
-<html><head><meta charset="utf-8"><title>Why complate?</title><link href="/assets/fonts.css" rel="stylesheet"><style>
-body {
-	margin: 0;
-	font-family: 'Sarabun', Trebuchet MS, Helvetica, sans-serif;
-	font-size: 1.125rem;
-}
-*, *::before, *::after {
-	box-sizing: border-box;
-}
-
-* + h2 {
-	margin-top: 5rem;
-}
-
-h2 {
-	font-size: 1.5rem;
-	font-weight: 300;
-	letter-spacing: 2px;
-	text-transform: uppercase;
-	color: hsl(205.7, 6.3%, 56.1%);
-}
-
-.container {
-	padding: 3rem 1rem;
-}
-.container > * {
-	margin-left: auto;
-	margin-right: auto;
-	max-width: 40rem;
-}
-
-.title-block {
-	background-color: #36454F;
-	color: #fff;
-}
-
-.title-block .container {
-	display: flex;
-	flex-wrap: wrap;
-	margin: 0 auto;
-	max-width: 40rem;
-}
-
-.title-block .title-area {
-	flex-basis: 5rem;
-	flex-grow: 1;
-}
-
-.title-block h1 {
-	font-size: 3rem;
-	font-family: 'Fira Code', monospace;
-}
-
-.title-block p {
-	font-size: 1.25rem;
-	color: #00f1ea;
-}
-
-.logo {
-	max-width: 200px;
-	max-height: 200px;
-	width: 100%;
-	margin: auto;
-}
-
-p {
-	line-height: 1.4;
-}
-
-blockquote {
-	margin-left: 1rem;
-	border-left: 3px solid #AAA;
-	padding-left: 0.5rem;
-}
-
-blockquote + p {
-	margin-left: 1rem;
-	font-size: 0.9rem;
-}
-
-img {
-	max-width: 100%;
-}
-</style><link href="/assets/code.css" rel="stylesheet"></head><body><div class="title-block"><div class="container"><img class="logo" src="complate.png"><div class="title-area"><h1>Why complate?</h1><p>Benefits of designing front-end code using a component abstraction</p><p></p></div></div></div><div class="container"><p>One of the main goals for complate is to create a templating engine which
+<html><head><meta charset="utf-8"><title>Why complate?</title><link href="/assets/fonts.css" rel="stylesheet"><link href="/assets/index.css" rel="stylesheet"></head><body><div class="title-block"><div class="container"><img class="logo" src="complate.png"><div class="title-area"><h1>Why complate?</h1><p>Benefits of designing front-end code using a component abstraction</p><p></p></div></div></div><div class="container"><p>One of the main goals for complate is to create a templating engine which
 allows us to write our front end using components. Why would we want to do
 that? Let’s look at some of the major benefits of a component based templating
 engine.</p>
