+++ views/components/code_example/_code_example.scss
@@ -1,0 +1,38 @@
+.code-example {
+	max-width: 60rem;
+	margin: 3rem auto;
+	display: flex;
+	flex-wrap: wrap;
+	border: 1px solid black;
+	background-color: #333;
+	color: #E5E5E5;
+	box-shadow: 0 10px 15px -3px rgba(0, 0, 0, 0.1), 0 4px 6px -2px rgba(0, 0, 0, 0.05);
+
+	pre {
+		margin: 0;
+	}
+
+	.input, .output {
+		flex-grow: 1;
+		margin: 0;
+		padding-top: 0.5rem;
+		overflow-x: auto;
+
+		&::before {
+			content: attr(aria-label);
+			display: block;
+			font-size: 0.875rem;
+			margin-left: 1rem;
+			font-family: 'Fira Code', monospace;
+			text-transform: uppercase;
+		}
+	}
+
+	.output {
+		background-color: white;
+		color: #555;
+	}
+	.output pre {
+		@extend %custom-prism-light-mode;		
+	}
+}
