+++ views/components/logo/_logo.scss
@@ -1,0 +1,6 @@
+.logo {
+	max-width: 200px;
+	max-height: 200px;
+	width: 100%;
+	margin: auto;
+}
