+++ views/components/logo/index.jsx
